refactor: extract drawable marquee from selection box
