+++ frontend/src/apps/slides/components/SlideContainer.vue
@@ -15,8 +15,9 @@
 				:isDragging
 				:rotationDelta="rotationDelta"
 				@mousedown="(e) => handleMouseDown(e)"
-				@setIsSelecting="(val) => (isSelecting = val)"
 			/>
+
+			<MarqueeOverlay v-if="!inReadonlyMode" @setIsSelecting="(val) => (isSelecting = val)" />
 
 			<SnapGuides
 				:ongoingInteraction="hasOngoingInteraction"
@@ -59,6 +60,7 @@ import { useResizeObserver } from '@vueuse/core'
 
 import SnapGuides from '@/apps/slides/components/SnapGuides.vue'
 import SelectionBox from '@/apps/slides/components/SelectionBox.vue'
+import MarqueeOverlay from '@/apps/slides/components/MarqueeOverlay.vue'
 import SlideElement from '@/apps/slides/components/SlideElement.vue'
 import DropTargetOverlay from '@/apps/slides/components/DropTargetOverlay.vue'
 import OverflowContentOverlay from '@/apps/slides/components/OverflowContentOverlay.vue'
